rename CompanionMenu -> CountryMenu, add translatable menu titles, add breadcrumb menu titles
