+++ src/main/resources/lang/en_GB.yml
@@ -106,6 +106,7 @@ menu-title:
   enter-player-name: "Enter player name"
   select-language: "Select Language"
   auto-detect-language: "Auto-Detect Language"
+  companion-select-continent: "Select A Continent"
 
 
 # -----------------------------------------------------
